% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_softpowers.r
\name{plot_softpowers}
\alias{plot_softpowers}
\title{Calculate and plot soft thresholding power diagnostics}
\usage{
\code{plot_softpowers(counts, powers = c(c(1:10), seq(from = 12, to=20, by=2)), verbose=5))}
}
\arguments{
\item{counts}{A matrix or data frame of counts, with genes in columns and samples in rows}

\item{powers}{A vector of soft thresholding powers; passed to \code{pickSoftThreshold}}

\item{verbose}{Integer level of verbosity; passed to \code{pickSoftThreshold}}
}
\description{
Calculate and plot diagnostics for a range of values of the soft thresholding power, as described
in the online WGCNA tutorials. It is a wrapper for WGCNA::pickSoftThreshold.
}
