% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signedKME2.R
\name{signedKME2}
\alias{signedKME2}
\title{Calculation of (signed) eigengene-based connectivity, without perturbing gene names.}
\usage{
\code{signedKME2(datExpr, datME, ...)}
}
\arguments{
\item{datExpr}{a data frame containing the gene expression data. Passed to signedKME.}

\item{dateME}{a data frame containing module eigengenes. Passed to signedKME.}
}
\value{
A data frame, as returned by \code{signedKME}, with any "_" in the gene names of \code{datExpr} retained as "_" in the result.
}
\description{
This is a wrapper for signedKME, with the additional step of reverting the gene names in the
result to the input gene names. signedKME converts all special characters in the gene names to
".", and this causes name mismatches downstream.
}
