% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdWGCNAmodules.r
\name{thresholdWGCNAmodules}
\alias{thresholdWGCNAmodules}
\title{Remove genes from modules if connectivity is below a specified threshold}
\usage{
\code{thresholdWGCNAmodules(gene_modules, threshold=0.25, color.no_mod="white",
  color_col="color_assigned", kME_col="kME_color_assigned")}
}
\arguments{
\item{gene_modules}{A data frame genes, with columns for the assigned module and connectivity to module.}

\item{threshold}{The connectivity threshold; genes with connectivity below this value will be removed. Defaults to 0.25.}

\item{color.no_mod}{The name for the new module, to which genes not meeting the threshold will be assigned.}

\item{color_col}{Optional, the name or number of the column in \code{gene_modules} which contains the module names.}

\item{kME_col}{Optional, the name or number of the column in \code{gene_modules} which contains the connectivity values.}
}
\value{
A new data frame with the same dimension and names as \code{gene_modules}, with 0 or more genes assigned to the new module.
}
\description{
Remove genes from WGCNA modules if connectivity to the module (kME) is below a specified threshold.
All such genes are assigned to a new module, which can be named by the user.
}
